+++ docs/API/utils.md
@@ -11,7 +11,3 @@
 ## calc_distances
 
 ::: embetter.utils.calc_distances
-
-## DifferenceClassifier
-
-::: embetter.utils.DifferenceClassifier
