+++ embetter/model/__init__.py
@@ -1,0 +1,4 @@
+from ._diff import DifferenceClassifier
+
+
+__all__ = ["DifferenceClassifier"]
